requirements.txt
Scrapy==0.22.2
Twisted==13.2.0
cffi==0.8.2
cryptography==0.2.2
cssselect==0.9.1
lxml==3.3.3
oauthlib==0.6.1
pyOpenSSL==0.14
pycparser==2.10
python-twitter==1.3.1
queuelib==1.1.1
requests==2.2.1
requests-oauthlib==0.4.0
scrapy-dblite==0.2.5
simplejson==3.3.3
six==1.6.1
w3lib==1.5
wsgiref==0.1.2
zope.interface==4.1.1
